Changes which children of notes are inactive.
Previously, all children of notes were inactive. This behavior is confusing and is not a good implementation of the original feature request, which was to have only notes that were children of notes to be inactive. This new implementation makes it so the original feature request is completed as intended.

Of note: I still don't have a working local host of the project, so I can't confirm this works 100%. I am running off the assumption that prop.ancestors[-1] returns the parent of a prop.